# gh-134173: Optimize concurrent.futures→asyncio state transfer with atomic snapshot