--- a/Lib/asyncio/futures.py
+++ b/Lib/asyncio/futures.py
@@ -351,36 +351,19 @@ def _set_concurrent_future_state(concurrent, source):
 def _copy_future_state(source, dest):
     """Internal helper to copy state from another Future.
 
-    The other Future may be a concurrent.futures.Future.
+    The other Future must be a concurrent.futures.Future.
     """
     if dest.cancelled():
         return
     assert not dest.done()
-
-    # Use _get_snapshot for futures that support it
-    if hasattr(source, '_get_snapshot'):
-        done, cancelled, result, exception = source._get_snapshot()
-        assert done
-        if cancelled:
-            dest.cancel()
-        elif exception is not None:
-            dest.set_exception(_convert_future_exc(exception))
-        else:
-            dest.set_result(result)
-        return
-
-    # Traditional fallback needs done check
-    assert source.done()
-    if source.cancelled():
+    done, cancelled, result, exception = source._get_snapshot()
+    assert done
+    if cancelled:
         dest.cancel()
+    elif exception is not None:
+        dest.set_exception(_convert_future_exc(exception))
     else:
-        exception = source.exception()
-        if exception is not None:
-            dest.set_exception(_convert_future_exc(exception))
-        else:
-            result = source.result()
-            dest.set_result(result)
-
+        dest.set_result(result)
 
 def _chain_future(source, destination):
     """Chain two futures so that when one completes, so does the other.