--- a/Misc/NEWS.d/next/Library/2025-05-18-07-25-15.gh-issue-134173.53oOoF.rst
+++ b/Misc/NEWS.d/next/Library/2025-05-18-07-25-15.gh-issue-134173.53oOoF.rst
@@ -1,4 +1,3 @@
-Add ``_get_snapshot()`` method to :class:`concurrent.futures.Future` to
-atomically retrieve all future state in a single lock acquisition. This speeds
-up :mod:`asyncio`'s ``_copy_future_state()`` by up to 4x when transferring state
-from :class:`concurrent.futures.Future` to :class:`asyncio.Future`. Patch by J. Nick Koston.
+Speed up :mod:`asyncio` performance of transferring state from thread
+pool :class:`concurrent.futures.Future` by up to 4.4x. Patch by J. Nick
+Koston.